Move functionality from main() to internal
